% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcc.R
\name{calc_cc}
\alias{calc_cc}
\title{Calculate Cross-Correlation}
\usage{
calc_cc(WxWy, na.rm = TRUE)
}
\arguments{
\item{WxWy}{A list containing the windows for two variables `x` and `y`.
Usually created by the `make_windows()` function.}

\item{na.rm}{A logical indicating whether to remove missing values from the
windows before calculating the cross-correlation (default = `TRUE`).}
}
\value{
A double between -1 and 1 indicating the cross-correlation or `NA`.
}
\description{
Calculate the cross-correlation of a pair of windows
}
