% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{calc_velocity_1d}
\alias{calc_velocity_1d}
\title{Calculate 1D Velocity}
\usage{
calc_velocity_1d(
  t,
  x,
  n = 1,
  method = c("central", "forward", "backward"),
  fill_edges = TRUE
)
}
\arguments{
\item{t}{A numeric vector representing time.}

\item{x}{A numeric vector of coordinates.}

\item{n}{An integer specifying the step size for the difference calculations. Default is 1.}

\item{method}{A character string specifying the method to use: "central", "forward", or "backward". Default is "central".}

\item{fill_edges}{A logical indicating whether to automatically use forward and backward differences to estimate velocities at the boundaries when `method = "central"`. Default is `TRUE`.}
}
\value{
A numeric vector of velocities the same length as the input vectors.
}
\description{
Calculates the velocity (rate of change including direction) along a single axis using finite difference methods.
}
