% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcc.R
\name{create_wcc_matrix}
\alias{create_wcc_matrix}
\title{Create wcc matrix}
\usage{
create_wcc_matrix(x, y, w_max, w_inc, tau_max, tau_inc)
}
\arguments{
\item{x}{A vector containing one time series (same length as `y`).}

\item{y}{A vector containing another time series (same length as `x`).}

\item{w_max}{Window size}

\item{w_inc}{Window increment}

\item{tau_max}{Largest lag size}

\item{tau_inc}{Lag increment}
}
\value{
A matrix
}
\description{
Create windowed cross-correlation matrix.
}
