% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leadership.R
\name{leadership_asymmetry}
\alias{leadership_asymmetry}
\title{Calculate Leadership Asymmetry Index}
\usage{
leadership_asymmetry(optima_obj, epoch_size = 10, min_valid = 3)
}
\arguments{
\item{optima_obj}{An object of class "wcc_optima".}

\item{epoch_size}{A positive integer specifying the number of windows to group
together to calculate the local asymmetry ratio.}

\item{min_valid}{A positive integer specifying the minimum number of valid (non-NA)
optima required in an epoch to compute the index.}
}
\value{
A data frame containing the rolling asymmetry index.
}
\description{
Computes a rolling index of leader-follower asymmetry from optimally picked lags.
A value of 1 indicates 'x' leads entirely, -1 indicates 'y' leads entirely,
and 0 indicates equal leading or simultaneous behavior.
}
