% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcc.R
\name{make_windows}
\alias{make_windows}
\title{Make windows}
\usage{
make_windows(x, y, i, tau, w_max)
}
\arguments{
\item{x}{A vector containing one time series (same length as `y`).}

\item{y}{A vector containing another time series (same length as `x`).}

\item{i}{A positive integer indicating the element in `x` and `y` to center
on.}

\item{tau}{An integer indicating the lag applied to `y`, i.e., the number of
elements to offset it relative to `x`.}

\item{w_max}{A positive integer indicating the size of the window, i.e., the
number of elements from `x` and `y` to include in the window.}
}
\value{
A list containing the windows `Wx` and `Wy`
}
\description{
Create two windows with the requested size and lagging.
}
