% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakpick.R
\name{pick_peaks}
\alias{pick_peaks}
\title{Find Peak Windowed Cross-Correlations}
\usage{
pick_peaks(wcc_obj, L_size, strict_monotonic = FALSE)
}
\arguments{
\item{wcc_obj}{An object of class "wcc_res" returned by `wcc()`.}

\item{L_size}{An odd integer specifying the size of the local search region.}

\item{strict_monotonic}{Logical indicating whether to strictly enforce
decreasing values on the flanks of the peak. (default = FALSE)}
}
\value{
A data frame containing the elapsed time indices, the peak lags,
  and the peak correlation values.
}
\description{
Find Peak Windowed Cross-Correlations
}
