% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leadership.R
\name{plot.bsync_lai}
\alias{plot.bsync_lai}
\title{Plot bsync_lai object}
\usage{
\method{plot}{bsync_lai}(x, time_step = 1, line_color = "#2166AC", smooth = FALSE, ...)
}
\arguments{
\item{x}{An object of class "bsync_lai".}

\item{time_step}{A numeric value specifying the duration of each index. Default is 1.}

\item{line_color}{Character string specifying the color of the LAI line. Default is "black".}

\item{smooth}{Logical indicating whether to add a loess smoothing line. Default is `FALSE`.}

\item{...}{Additional arguments (not used).}
}
\description{
Plot bsync_lai object
}
