% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdtw_plot.R
\name{plot.wdtw_res}
\alias{plot.wdtw_res}
\title{Plot wdtw_res object}
\usage{
\method{plot}{wdtw_res}(x, time_step = 1, show_zero_lag = TRUE, zero_line_color = "black", ...)
}
\arguments{
\item{x}{An object of class "wdtw_res".}

\item{time_step}{A numeric value specifying the duration of each index.
If not 1, axes will be converted from raw indices to time units. Default is 1.}

\item{show_zero_lag}{Logical indicating whether to draw a vertical line at lag = 0. Default is `TRUE`.}

\item{zero_line_color}{Character string specifying the color of the zero-lag line. Default is "black".}

\item{...}{Additional arguments (not used).}
}
\description{
Plot wdtw_res object
}
