% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgranger_plot.R
\name{plot.wgranger_res}
\alias{plot.wgranger_res}
\title{Plot wgranger_res object}
\usage{
\method{plot}{wgranger_res}(x, time_step = 1, metric = c("F", "p"), smooth = FALSE, ...)
}
\arguments{
\item{x}{An object of class "wgranger_res".}

\item{time_step}{A numeric value specifying the duration of each index. Default is 1.}

\item{metric}{Character string specifying which metric to plot. Options are
`"F"` for F-Statistics or `"p"` for negative log10 p-values. Default is `"F"`.}

\item{smooth}{Logical indicating whether to apply loess smoothing to the lines. Default is `FALSE`.}

\item{...}{Additional arguments (not used).}
}
\description{
Plot wgranger_res object
}
