% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optima_plot.R
\name{plot_optima_overlay}
\alias{plot_optima_overlay}
\title{Plot Surface with Optima Overlay}
\usage{
plot_optima_overlay(
  surface_obj,
  optima_df,
  time_step = 1,
  line_color = "black",
  point_fill = "black",
  point_stroke = "white",
  show_zero_lag = TRUE,
  zero_line_color = "black",
  ...
)
}
\arguments{
\item{surface_obj}{An object of class "wcc_res" or "wdtw_res".}

\item{optima_df}{A data frame of class "wcc_optima" or "wdtw_optima".}

\item{time_step}{A numeric value specifying the duration of each index. Default is 1.}

\item{line_color}{Character string specifying the color of the connecting line. Default is "black".}

\item{point_fill}{Character string specifying the inner fill of the optima points. Default is "black".}

\item{point_stroke}{Character string specifying the outer outline of the optima points. Default is "white".}

\item{show_zero_lag}{Logical indicating whether to draw a vertical line at lag = 0. Default is `TRUE`.}

\item{zero_line_color}{Character string specifying the color of the zero-lag line. Default is "black".}

\item{...}{Additional arguments passed to the underlying plot method (e.g., custom colors for WCC).}
}
\description{
Plot Surface with Optima Overlay
}
