% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcc_plot.R
\name{plot_peaks_overlay}
\alias{plot_peaks_overlay}
\title{Plot Windowed Cross-Correlation with Peak Overlay}
\usage{
plot_peaks_overlay(wcc_obj, peaks_df)
}
\description{
Plot Windowed Cross-Correlation with Peak Overlay
}
