% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leadership.R
\name{print.bsync_lai}
\alias{print.bsync_lai}
\title{Print method for bsync_lai objects}
\usage{
\method{print}{bsync_lai}(x, n = 5, ...)
}
\arguments{
\item{x}{An object of class "bsync_lai".}

\item{n}{An integer specifying how many rows to print. Default is 5.}

\item{...}{Additional arguments (not used).}
}
\description{
Print method for bsync_lai objects
}
