% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optima.R
\name{print.wcc_optima}
\alias{print.wcc_optima}
\title{Print method for wcc_optima objects}
\usage{
\method{print}{wcc_optima}(x, n = 5, ...)
}
\arguments{
\item{x}{An object of class "wcc_optima".}

\item{n}{An integer specifying how many rows to print. Default is 5.}

\item{...}{Additional arguments (not used).}
}
\description{
Print method for wcc_optima objects
}
