% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakpick.R
\name{print.wcc_peaks}
\alias{print.wcc_peaks}
\title{Print method for wcc_peaks objects}
\usage{
\method{print}{wcc_peaks}(x, n = 5, ...)
}
\arguments{
\item{x}{An object of class "wcc_peaks".}

\item{n}{An integer specifying how many rows of peaks to print. Default is 5.}

\item{...}{Additional arguments (not used).}
}
\description{
Print method for wcc_peaks objects
}
