% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcc.R
\name{print.wcc_res}
\alias{print.wcc_res}
\title{Print method for wcc_res objects}
\usage{
\method{print}{wcc_res}(x, ...)
}
\arguments{
\item{x}{An object of class "wcc_res".}

\item{...}{Additional arguments (not used).}
}
\description{
Print method for wcc_res objects
}
