% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surrogate.R
\name{print.wcc_surr}
\alias{print.wcc_surr}
\title{Print method for wcc_surr objects}
\usage{
\method{print}{wcc_surr}(x, ...)
}
\arguments{
\item{x}{An object of class "wcc_surr".}

\item{...}{Additional arguments (not used).}
}
\description{
Print method for wcc_surr objects
}
