% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optima.R
\name{print.wdtw_optima}
\alias{print.wdtw_optima}
\title{Print method for wdtw_optima objects}
\usage{
\method{print}{wdtw_optima}(x, n = 5, ...)
}
\arguments{
\item{x}{An object of class "wdtw_optima".}

\item{n}{An integer specifying how many rows to print. Default is 5.}

\item{...}{Additional arguments (not used).}
}
\description{
Print method for wdtw_optima objects
}
