% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdtw.R
\name{print.wdtw_res}
\alias{print.wdtw_res}
\title{Print method for wdtw_res objects}
\usage{
\method{print}{wdtw_res}(x, ...)
}
\arguments{
\item{x}{An object of class "wdtw_res".}

\item{...}{Additional arguments (not used).}
}
\description{
Print method for wdtw_res objects
}
