% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgranger.R
\name{print.wgranger_res}
\alias{print.wgranger_res}
\title{Print method for wgranger_res objects}
\usage{
\method{print}{wgranger_res}(x, ...)
}
\arguments{
\item{x}{An object of class "wgranger_res".}

\item{...}{Additional arguments (not used).}
}
\description{
Print method for wgranger_res objects
}
