% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optima.R
\name{summary.wcc_optima}
\alias{summary.wcc_optima}
\title{Summary method for wcc_optima objects}
\usage{
\method{summary}{wcc_optima}(object, ...)
}
\arguments{
\item{object}{An object of class "wcc_optima".}

\item{...}{Additional arguments (not used).}
}
\description{
Summary method for wcc_optima objects
}
