% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcc.R
\name{summary.wcc_res}
\alias{summary.wcc_res}
\title{Summary method for wcc_res objects}
\usage{
\method{summary}{wcc_res}(object, ...)
}
\arguments{
\item{object}{An object of class "wcc_res".}

\item{...}{Additional arguments (not used).}
}
\description{
Summary method for wcc_res objects
}
