% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optima.R
\name{summary.wdtw_optima}
\alias{summary.wdtw_optima}
\title{Summary method for wdtw_optima objects}
\usage{
\method{summary}{wdtw_optima}(object, ...)
}
\arguments{
\item{object}{An object of class "wdtw_optima".}

\item{...}{Additional arguments (not used).}
}
\description{
Summary method for wdtw_optima objects
}
