% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgranger.R
\name{summary.wgranger_res}
\alias{summary.wgranger_res}
\title{Summary method for wgranger_res objects}
\usage{
\method{summary}{wgranger_res}(object, ...)
}
\arguments{
\item{object}{An object of class "wgranger_res".}

\item{...}{Additional arguments (not used).}
}
\description{
Summary method for wgranger_res objects
}
