% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{v_x}
\alias{v_x}
\title{Calculate 1D Velocity Magnitude (X-Axis)}
\usage{
v_x(t, x, n = 1)
}
\arguments{
\item{t}{A numeric vector representing time.}

\item{x}{A numeric vector of x-coordinates.}

\item{n}{An integer specifying the step size for the lag and lead calculations. Default is 1.}
}
\value{
A numeric vector of velocities. The first and last `n` elements will be `NA`.
}
\description{
Calculates the absolute velocity (speed) along the x-axis using a central difference method.
}
