% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{v_xyz}
\alias{v_xyz}
\title{Calculate 3D Velocity Magnitude (Speed)}
\usage{
v_xyz(t, x, y, z, n = 1)
}
\arguments{
\item{t}{A numeric vector representing time.}

\item{x}{A numeric vector of x-coordinates.}

\item{y}{A numeric vector of y-coordinates.}

\item{z}{A numeric vector of z-coordinates.}

\item{n}{An integer specifying the step size for the lag and lead calculations. Default is 1.}
}
\value{
A numeric vector of velocities. The first and last `n` elements will be `NA`.
}
\description{
Calculates the magnitude of the 3D velocity vector (speed) using a central difference method.
}
