% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{v_y}
\alias{v_y}
\title{Calculate 1D Velocity Magnitude (Y-Axis)}
\usage{
v_y(t, y, n = 1)
}
\arguments{
\item{t}{A numeric vector representing time.}

\item{y}{A numeric vector of y-coordinates.}

\item{n}{An integer specifying the step size for the lag and lead calculations. Default is 1.}
}
\value{
A numeric vector of velocities. The first and last `n` elements will be `NA`.
}
\description{
Calculates the absolute velocity (speed) along the y-axis using a central difference method.
}
