% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{v_z}
\alias{v_z}
\title{Calculate 1D Velocity Magnitude (Z-Axis)}
\usage{
v_z(t, z, n = 1)
}
\arguments{
\item{t}{A numeric vector representing time.}

\item{z}{A numeric vector of z-coordinates.}

\item{n}{An integer specifying the step size for the lag and lead calculations. Default is 1.}
}
\value{
A numeric vector of velocities. The first and last `n` elements will be `NA`.
}
\description{
Calculates the absolute velocity (speed) along the z-axis using a central difference method.
}
