% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgranger.R
\name{wgranger}
\alias{wgranger}
\title{Windowed Granger Causality}
\usage{
wgranger(x, y, time = NULL, window_size, ar_order = 1, window_increment = 1)
}
\arguments{
\item{x}{A numeric vector containing a time series (same length as `y`).}

\item{y}{A numeric vector containing a time series (same length as `x`).}

\item{time}{An optional numeric vector representing the timestamps for the data.}

\item{window_size}{A positive integer indicating the size of each window.}

\item{ar_order}{A positive integer specifying the Autoregressive (AR) order.
This represents the maximum number of lags included in the prediction model.}

\item{window_increment}{A positive integer indicating the window shift increment. (default = `1`)}
}
\value{
A list object of class "wgranger_res".
}
\description{
Conduct a rolling windowed Granger Causality analysis to determine dynamic
leader-follower relationships between two continuous time series.
}
