% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakpick.R
\name{pick_peaks}
\alias{pick_peaks}
\title{Find Peak (or Valley) Windowed Cross-Correlations}
\usage{
pick_peaks(wcc_obj, L_size, strict_monotonic = FALSE, find_min = FALSE)
}
\arguments{
\item{wcc_obj}{An object of class "wcc_res" or "wdtw_res".}

\item{L_size}{An odd integer specifying the size of the local search region.}

\item{strict_monotonic}{Logical indicating whether to strictly enforce
monotonic flanks around the extremum. (default = FALSE)}

\item{find_min}{Logical indicating whether to search for local minima instead
of local maxima. Required for distance metrics like DTW. (default = FALSE)}
}
\value{
A data frame of class "wcc_peaks".
}
\description{
Find Peak (or Valley) Windowed Cross-Correlations
}
