% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leadership.R
\name{plot.bsync_lai}
\alias{plot.bsync_lai}
\title{Plot bsync_lai object}
\usage{
\method{plot}{bsync_lai}(
  x,
  time_step = 1,
  x_label = NULL,
  line_color = "#2166AC",
  smooth = FALSE,
  smooth_color = "#B2182B",
  ...
)
}
\arguments{
\item{x}{An object of class "bsync_lai".}

\item{time_step}{A numeric value specifying the duration of each index. Default is 1.}

\item{x_label}{An optional character string to override the default x-axis label.}

\item{line_color}{Character string specifying the color of the LAI line. Default is "#2166AC".}

\item{smooth}{Logical indicating whether to add a loess smoothing line. Default is `FALSE`.}

\item{smooth_color}{Character string specifying the color of the smoothing line. Default is "#B2182B".}

\item{...}{Additional arguments (not used).}
}
\description{
Plot bsync_lai object
}
