% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcc_plot.R
\name{plot.wcc_res}
\alias{plot.wcc_res}
\title{Plot wcc_res object}
\usage{
\method{plot}{wcc_res}(x, ...)
}
\arguments{
\item{x}{An object of class "wcc_res".}

\item{...}{Additional arguments (not used).}
}
\description{
Plot wcc_res object
}
