% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcc_plot.R
\name{plot_peaks_overlay}
\alias{plot_peaks_overlay}
\title{Plot Windowed Cross-Correlation with Peak Overlay}
\usage{
plot_peaks_overlay(
  wcc_obj,
  peaks_df,
  time_step = 1,
  line_color = "black",
  point_fill = "black",
  point_stroke = "white",
  color_low = "#B2182B",
  color_mid = "#F7F7F7",
  color_high = "#2166AC"
)
}
\arguments{
\item{wcc_obj}{An object of class "wcc_res" containing the windowed cross-correlation results.}

\item{peaks_df}{A data frame of class "wcc_peaks" containing the peaks to overlay.}

\item{time_step}{A numeric value specifying the duration of each index. Default is 1.}

\item{line_color}{Character string specifying the color of the connecting line. Default is "black".}

\item{point_fill}{Character string specifying the inner fill of the peak points. Default is "black".}

\item{point_stroke}{Character string specifying the outer outline of the peak points. Default is "white".}

\item{color_low}{Character string specifying the heatmap color for a correlation of -1. Default is "#B2182B".}

\item{color_mid}{Character string specifying the heatmap color for a correlation of 0. Default is "#F7F7F7".}

\item{color_high}{Character string specifying the heatmap color for a correlation of 1. Default is "#2166AC".}
}
\description{
Plot Windowed Cross-Correlation with Peak Overlay
}
