% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcc_plot.R
\name{plot_peaks_overlay}
\alias{plot_peaks_overlay}
\title{Plot Windowed Cross-Correlation with Peak Overlay}
\usage{
plot_peaks_overlay(wcc_obj, peaks_df)
}
\arguments{
\item{wcc_obj}{An object of class "wcc_res" containing the windowed cross-correlation results.}

\item{peaks_df}{A data frame containing the peaks to overlay.}
}
\description{
Plot Windowed Cross-Correlation with Peak Overlay
}
