% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcc-package.R
\docType{package}
\name{wcc-package}
\alias{wcc-package}
\title{wcc: Windowed Cross-Correlation Analysis}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Jeffrey Girard \email{me@jmgirard.com}

Authors:
\itemize{
  \item Jeffrey Girard \email{me@jmgirard.com}
}

}
\keyword{internal}
