% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdtw.R
\name{wdtw}
\alias{wdtw}
\title{Windowed Dynamic Time Warping}
\usage{
wdtw(
  x,
  y,
  window_size,
  lag_max,
  window_increment = 1,
  lag_increment = 1,
  scale_data = TRUE
)
}
\arguments{
\item{x}{A numeric vector containing a time series (same length as `y`).}

\item{y}{A numeric vector containing a time series (same length as `x`).}

\item{window_size}{A positive integer indicating the size of each window.}

\item{lag_max}{A positive integer indicating the maximum lag to try.}

\item{window_increment}{A positive integer indicating the window shift increment. (default = `1`)}

\item{lag_increment}{A positive integer indicating the lag shift increment. (default = `1`)}

\item{scale_data}{A logical indicating whether to z-score standardize both
time series prior to calculation. Highly recommended for DTW. (default = `TRUE`)}
}
\value{
A list object of class "wdtw_res".
}
\description{
Conduct a windowed dynamic time warping (WDTW) analysis to find the optimal
alignment distance between sliding windows of two time series.
}
